refactor(inventory): split core.rs by handler (#244)
core.rs was 701 lines (just over the 700-line cap). The three large
handlers (remove-by-instance / use-item / remove-by-type) are
independent transactions; split each into its own file:

  core/mod.rs            — INVENTORY_ITEM_SELECT canonical query,
                           InventoryRow / InventoryInstanceRow /
                           InventoryInstanceWithIdRow,
                           send_full_inventory_update,
                           send_on_remove_item, module decls + re-exports
  core/remove_instance.rs — handle_remove_inventory_item (by item_id)
  core/use_instance.rs    — handle_use_inventory_item (fires OnItemUse)
  core/remove_by_type.rs  — handle_remove_inventory_item_by_type (by type_id)

Public surface preserved: parent `inventory/mod.rs`'s
`pub use core::{handle_remove_inventory_item, ...,
send_full_inventory_update}` block resolves through the new mod.rs
re-exports unchanged. The `pub(crate) const INVENTORY_ITEM_SELECT`
stays accessible at `inventory::core::INVENTORY_ITEM_SELECT` so the
SQL drift-guard test in `player_load/core.rs` keeps working.

Tests pass; clippy clean; cargo fmt applied.
